Artist's Statement
When I have been tasked with creating something, either a poster, brochure, website or game, I am quite flexible as an artist and don’t stick to only one particular style. As someone who has studied in the field of graphic design, I have learned that the style of the creation you made often reflects the client’s vision and must appeal to the targeted audience. When straying away from these restraints, I often enjoy exploring different styles and concepts that are fun, creative or even odd. I enjoy having this creative freedom since I hadn’t had the opportunity in past courses and this project is one that allowed this for me.
The task given for this project was to familiarize ourselves with the story of Sisyphus through reading articles and watching the 1974 animation done by Marcell Jankovic. Upon looking at these sources, we were to come up with a game in which reflects the story of Sisyphus in a metaphorical way. Sisyphus is a figure in Greek mythology, who was punished for eternity by having to push a boulder up a cliff but being unable to ever reach the top as the boulder rolled back down just as he was about to. He must keep experiencing this vicious cycle with no escape. The idea of my game was to have a similar concept of having to move something from point A to point B and it resetting to point A when it reached point B. The idea that came to mind is based on a connection I made watching the Sisyphus animation which was the resemblance between a boulder and a meatball. I then came up with the game being the user having to drag a fork to a meatball, which would then be “picked up”. The meatball on the fork would then be dragged to a mouth but once it reaches it, the empty fork would reset to its initial position and the meatball be “spit” out back onto the plate, along with spitting out sound effect. This would then be the whole cycle of the game. 
The way in which I created this game was using HTML, CSS, and jQuery UI. I started by first collecting the images I wanted to use, putting them into HTML and used CSS to position everything. Then using jQuery UI, I made the different attributes draggable (fork and fork on meatball) or droppable (meatball and mouth). I then made the objects interact with one another, causing items to disappear or appear (ex: the fork overlaps the meatball and the meatball on fork image appears). Also, used animation to portray the meatball being spit out by having it move from the top right to bottom left. I also used draggable delay with was used with the fork with meatball. It has a slight delay at the start to portray resistance. Lastly, along with the meatball animation, I added an audio effect of spitting to be triggered when the mouth is overlapped with the meatball on the fork. 
I interpret Sisyphus as a story of perseverance and repetition, which I believe is well reflected in the game I created. Although the vast difference such as my fun approach to the story it is also an online game in which the user can give on and close the window at any point, unlike Sisyphus who was trapped in the same routine for eternity.